fixed code beautiful
