+++ includes/footer.php
@@ -1,0 +1,19 @@
+<footer class="footerbot">
+        
+            <div class="ui list" id="footer-links">
+                <a href="#vanligafrågor" class="item">Vanliga frågor</a>
+                <a href="#leveransvillkor" class="item">Leveransvillkor</a>
+                <a href="#policy" class="item">Integritetspolicy</a>
+        </div>
+        <div class="footer-images">
+            <!-- <img class="payment-method" src="assets/klarna.png">
+            <img class="payment-method" src="assets/payment-cards.png">
+            <img class="payment-method" src="assets/swish.png"> -->
+            <div class="safe-ecommerce">
+                <img class="trygg-logo" src="assets/trygg-ehandel.png">
+                <img class="certifierad-logo" src="assets/certifierad-ehandel.png">
+            </div>
+            <img class="payment-method" src="assets/payment-method.png">
+
+        </div>
+</footer>
