+++ includes/loginForm.php
@@ -1,0 +1,42 @@
+<div class="ui modal test">
+<div class="ui middle aligned center aligned grid">
+<div class="column">
+<h2 class="ui red header">
+<div class="content">
+DALAMAT
+</div>
+</h2>
+
+<form class="ui large form login-form" method="POST" data-ajax="false">
+<div class="ui segment">
+<div class="field">
+<div class="ui left icon input">
+    <i class="user icon"></i>
+    <input type="text" name="username" id="username" placeholder="Användarnamn">
+</div>
+</div>
+<div class="field">
+<div class="ui left icon input">
+    <i class="lock icon"></i>
+    <input type="password" name="password" id="password" placeholder="Password">
+</div>
+</div>
+<div>
+<input type="submit" name="Login" value="Logga in" class="ui fluid large green submit button" id="buttonLog" />
+</div>
+
+</div>
+
+<div class="ui error message"></div>
+<div class="error-message"></div>
+</form>
+<!-- Redirect too register.php with message -->
+<div class="ui message">
+Ny hos oss? <a href="register.php">Registera dig</a>
+</div>
+</div>
+
+</div>
+
+</div>
+</div>
